+++ Tests/Unit/Utility/SqlInjectionUtilityTest.php
@@ -44,4 +44,109 @@ namespace Slavlee\Waf\Tests\Unit\Utility;
             SqlInjectionUtility::scanString($stringToScan)
         );
     }
+
+    /**
+     * @test
+     * Test 2 if Utility can detect "OR 1=1" SQL queries
+     */
+    public function denyOrZeroEqualsZeroString()
+    {
+        $stringToScan = '105 OR 1=1';
+        $expectedResult = false;
+
+        $this->assertEquals(
+            $expectedResult,
+            SqlInjectionUtility::scanString($stringToScan)
+        );
+    }
+
+    /**
+     * @test
+     * Test 2 if Utility can detect 'OR "x"="x"' SQL queries
+     */
+    public function denyOrXEqualsXString()
+    {
+        $stringToScan = '105 OR "x"="x"';
+        $expectedResult = false;
+
+        $this->assertEquals(
+            $expectedResult,
+            SqlInjectionUtility::scanString($stringToScan)
+        );
+    }
+
+    /**
+     * @test
+     * Test 2 if Utility can detect "OR 'x'='x'" SQL queries
+     */
+    public function denyOrXEqualsXV2String()
+    {
+        $stringToScan = '105 OR \'x\'=\'x\'';
+        $expectedResult = false;
+
+        $this->assertEquals(
+            $expectedResult,
+            SqlInjectionUtility::scanString($stringToScan)
+        );
+    }
+
+    /**
+     * @test
+     * Test 2 if Utility can detect "OR 1=1" SQL queries
+     */
+    public function denyAndZeroEqualsZeroString()
+    {
+        $stringToScan = '105 AND 1=1';
+        $expectedResult = false;
+
+        $this->assertEquals(
+            $expectedResult,
+            SqlInjectionUtility::scanString($stringToScan)
+        );
+    }
+
+    /**
+     * @test
+     * Test 2 if Utility can detect 'OR "x"="x"' SQL queries
+     */
+    public function denyAndXEqualsXString()
+    {
+        $stringToScan = '105 AND "x"="x"';
+        $expectedResult = false;
+
+        $this->assertEquals(
+            $expectedResult,
+            SqlInjectionUtility::scanString($stringToScan)
+        );
+    }
+
+    /**
+     * @test
+     * Test if Utility can detect 'database()' SQL queries
+     */
+    public function denyDatabaseFuncString()
+    {
+        $stringToScan = 'database()';
+        $expectedResult = false;
+
+        $this->assertEquals(
+            $expectedResult,
+            SqlInjectionUtility::scanString($stringToScan)
+        );
+    }
+
+    /**
+     * @test
+     * Test if Utility can detect 'INFORMATION_SCHEMA' SQL queries
+     */
+    public function denyInformationSchemaString()
+    {
+        $stringToScan = 'INFORMATION_SCHEMA.TABLES';
+        $expectedResult = false;
+
+        $this->assertEquals(
+            $expectedResult,
+            SqlInjectionUtility::scanString($stringToScan)
+        );
+    }
  }